+++ website_sale_preset_carts/__manifest__.py
@@ -23,8 +23,6 @@
         "demo/demo.xml",  # fixme
     ],
     "data": [
-        "data/cron.xml",
-        "data/mail_template.xml",
         "security/ir.model.access.csv",
         "views/auth_signup_template.xml",
         "views/preset_cart.xml",
