+++ website_sale_preset_carts/views/res_config_settings.xml
@@ -35,26 +35,4 @@
           </xpath>
         </field>
     </record>
-    <record id="res_config_settings_view_form_distribution_circuits_website_sale" model="ir.ui.view">
-        <field name="name">
-          res.config.settings.view.form.distribution.circuits.website.sale
-        </field>
-        <field name="model">res.config.settings</field>
-        <field name="inherit_id" ref="website_sale.res_config_settings_view_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//div[@id='checkout_mail_setting']" position="before">
-                <div class="col-xs-12 col-md-6 o_setting_box" id="send_mail_to_supervisor">
-                    <div class="o_setting_left_pane">
-                        <field name="send_mail_to_supervisor"/>
-                    </div>
-                    <div class="o_setting_right_pane">
-                        <label for="send_mail_to_supervisor"/>
-                        <div class="text-muted">
-                            Send notification emails to Time frame supervisor on opening and closing.
-                        </div>
-                    </div>
-                </div>
-          </xpath>
-        </field>
-      </record>
 </odoo>
