+++ website_sale_time_frame_auto/__init__.py
@@ -1,0 +1,1 @@
+from . import models
