+++ website_sale_time_frame_auto/models/__init__.py
@@ -1,0 +1,2 @@
+from . import res_config_settings
+from . import time_frame
