+++ website_sale_time_frame_auto/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_time_frame
