+++ website_sale_time_frame_auto/views/res_config_settings.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2018 Coop IT Easy SCRLfs
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+    <record id="res_config_settings_view_form_website_sale_auto_open_close" model="ir.ui.view">
+        <field name="name">
+          res.config.settings.view.form.website.sale.auto.open.close
+        </field>
+        <field name="model">res.config.settings</field>
+        <field name="inherit_id" ref="website_sale.res_config_settings_view_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//div[@id='checkout_mail_setting']" position="before">
+                <div class="col-xs-12 col-md-6 o_setting_box" id="send_mail_to_supervisor">
+                    <div class="o_setting_left_pane">
+                        <field name="send_mail_to_supervisor"/>
+                    </div>
+                    <div class="o_setting_right_pane">
+                        <label for="send_mail_to_supervisor"/>
+                        <div class="text-muted">
+                            Send notification emails to Time frame supervisor on opening and closing.
+                        </div>
+                    </div>
+                </div>
+          </xpath>
+        </field>
+      </record>
+</odoo>
